🚧 依旧数据库施工中

+++ core/models/tag.go
@@ -12,7 +12,7 @@ type Tag struct {
 	UpdatedAt time.Time      `gorm:"column:updated_at;type:timestamp;not null;default:CURRENT_TIMESTAMP;comment:更新时间"`
 	DeletedAt gorm.DeletedAt `gorm:"column:deleted_at;type:timestamp;index;comment:删除时间（软删除）"`
 
-	TagID int64  `gorm:"column:tag_id;type:bigint;uniqueIndex:uk_tag_id;not null;comment:标签ID（唯一）" json:"tag_id"`       // 标签名唯一
+	TagID uint32 `gorm:"column:tag_id;type:bigint;uniqueIndex:uk_tag_id;not null;comment:标签ID（唯一）" json:"tag_id"`       // 标签名唯一
 	Name  string `gorm:"column:name;type:varchar(50);uniqueIndex:uk_tag_name;not null;comment:标签名（唯一）" json:"name"`     // 标签名唯一
 	Slug  string `gorm:"column:slug;type:varchar(50);uniqueIndex:uk_tag_slug;not null;comment:URL友好标识（唯一）" json:"slug"` // URL友好标识
